+++ pytsite/package_info/__init__.py
@@ -2,7 +2,7 @@
 """
 from . import _error as error
 from ._api import resolve_path, parse_json, data, requires, requires_packages, requires_plugins, name, description, \
-    version, url, check_requirements
+    version, url
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'