+++ pytsite/setup/_console_command.py
@@ -2,7 +2,8 @@
 """
 from datetime import datetime as _datetime
 from os import path as _path, makedirs as _makedirs
-from pytsite import console as _console, reg as _reg, events as _events, lang as _lang
+from pytsite import console as _console, reg as _reg, events as _events, lang as _lang, util as _util, \
+    package_info as _package_info, theme as _theme
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -36,6 +37,15 @@ class Setup(_console.Command):
         lock_path = _reg.get('paths.setup.lock')
         if _path.exists(lock_path) and not self.opt('force'):
             raise _console.error.Error(_lang.t('pytsite.setup@setup_is_already_completed'))
+
+        # Install required pip packages by application and theme
+        for pkg_name in ['app', _theme.get().package_name]:
+            for pip_pkg_spec in _package_info.requires_packages(pkg_name):
+                _console.print_info(_lang.t('pytsite.setup@installing_pip_package', {'package': pip_pkg_spec}))
+                try:
+                    _util.install_pip_package(pip_pkg_spec)
+                except _util.error.PipPackageInstallError as e:
+                    raise _console.error.Error(e)
 
         _events.fire('pytsite.setup')