+++ pytsite/update/__init__.py
@@ -1,7 +1,7 @@
 """PytSite Update
 """
 # API functions
-from ._api import on_update, on_update_after
+from ._api import on_update_stage_1, on_update_stage_2, on_update, on_update_after
 
 from pytsite import console as _console, lang as _lang
 from . import _console_command