+++ pytsite/update/_api.py
@@ -8,6 +8,18 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
+def on_update_stage_1(handler: _Callable[[str], None], priority: int = 0):
+    """Shortcut.
+    """
+    _events.listen('pytsite.update.stage_1', handler, priority)
+
+
+def on_update_stage_2(handler: _Callable[[str], None], priority: int = 0):
+    """Shortcut.
+    """
+    _events.listen('pytsite.update.stage_2', handler, priority)
+
+
 def on_update(handler: _Callable[[str], None], priority: int = 0):
     """Shortcut.
     """